Add files via upload
This is complete Event Registration Confirmation system that i have built, it takes data from a Google Form and automatically creates a unique Registration ID and QR code for every participant. The system stores all records in Google Sheets, saves the QR code in Google Drive, and sends a confirmation email with event details and the QR code. This makes check-in at the venue fast and professional, and also helps build accurate attendance data for future marketing.